working version with excluded apps

+++ android/app/src/main/kotlin/org/getlantern/lantern/apps/AppData.kt
@@ -2,7 +2,7 @@ package org.getlantern.lantern.apps
 
 data class AppData(
     val packageName: String,
-    val iconRes: Int,
+    val iconRes: Long,
     val name: String,
     val isSystemApp: Boolean = false
 )
